api/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.103.2
pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
